refactor(login): Put relevant code into view model and double checked that the friends list query is valid. It is valid, but we need a list to play with.
